A search engine for recipes by ingredients
Due to the recent outbreak of COVID-19, there are many who are experiences trouble getting their usual groceries. Because of this I've created a search engine. All you have to do is input ingredients you want to use, and it will show you different options of meals you can make.